refactor: change Header to Navbar and update styling
