+++ frontend/src/pages/Help.jsx
@@ -1,9 +1,9 @@
-import Header from "../components/Header";
+import Navbar from "../components/Navbar";
 
 function Help() {
     return (
         <div>
-            <Header />
+            <Navbar />
             <h1>Help Page</h1>
         </div>
     );
