+++ frontend/src/pages/Profile.jsx
@@ -1,9 +1,9 @@
-import Header from "../components/Header";
+import Navbar from "../components/Navbar";
 
 function Profile() {
     return (
         <div>
-            <Header />
+            <Navbar />
             <h1>Profile Page</h1>
         </div>
     );
